Move theatre listener to AutoTheatreMode module
