+++ src/modules/auto_theater_mode/index.js
@@ -1,6 +1,10 @@
 const debug = require('../../utils/debug');
 const settings = require('../../settings');
 const watcher = require('../../watcher');
+const storage = require('../../storage');
+
+let tempToggledDarkMode = false;
+
 
 class AutoTheaterModeModule {
     constructor() {
@@ -15,18 +19,28 @@ class AutoTheaterModeModule {
     }
 
     load() {
+        this.attachTheatreListener();
+
         if (settings.get('autoTheatreMode') === false) return;
 
         try {
-            const container = App.__container__;
-            const routeName = container.lookup('controller:application').get('currentRouteName');
-            if (routeName !== 'channel.index.index' && ['videos', 'vod'].includes(routeName)) return;
-            if (container.lookup('service:persistentPlayer').playerComponent.player.theatre) return;
-
+            if (App.__container__.lookup('service:persistentPlayer').playerComponent.player.theatre) return;
             window.Mousetrap.trigger('alt+t');
         } catch (e) {
             debug.log('Error toggling theater mode: ', e);
         }
+    }
+
+    attachTheatreListener() {
+        const playerService = App.__container__.lookup('service:persistentPlayer');
+        if (!playerService || !playerService.playerComponent || !playerService.playerComponent.player) return;
+        playerService.playerComponent.player.addEventListener('theatrechange', state => {
+            if (settings.get('darkenedMode') === true && !tempToggledDarkMode) return;
+            tempToggledDarkMode = !tempToggledDarkMode;
+            settings.set('darkenedMode', state);
+            // Turn darkenedMode setting off again if needed but without emit
+            if (state) storage.set('darkenedMode', false, undefined, false, false);
+        });
     }
 }
